Refactored <ListingForm> into it's own component.
