+++ components/forms/index.js
@@ -1,0 +1,2 @@
+export { default as ListingForm } from './listing';
+
