+++ components/index.js
@@ -1,2 +1,3 @@
+export * from './forms';
 export { default as Header } from './header';
