+++ pages/index.js
@@ -1,35 +1,16 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import Router from 'next/router';
-import { connect } from 'react-redux';
 
-import { apply } from 'reputable/dist/redux/action-creators/registry';
-import { getParticipants, getRegistryAddress } from 'reputable/dist/redux/selectors';
+import { Header, ListingForm } from '../components';
 
-import { Header } from '../components';
-
-class Page extends React.Component {
+class IndexPage extends React.Component {
   constructor() {
     super();
 
-    this.handleSubmit = this.handleSubmit.bind(this);
+    this.handleAfterSubmit = this.handleAfterSubmit.bind(this);
   }
 
-  async handleSubmit(e) {
-    e.preventDefault();
-
-    const {
-      registryAddress,
-      participants,
-      submitApplication,
-    } = this.props;
-
-    const userAddress = participants.length ? participants[1].address : '';
-
-    const formData = new FormData(e.target);
-    const value = formData.get('university');
-
-    await submitApplication(registryAddress, value, userAddress, 100);
+  handleAfterSubmit() {
     Router.push('/listings');
   }
 
@@ -37,39 +18,11 @@ class Page extends React.Component {
     return (
       <div>
         <Header />
-        <h1>Enter the name of your university</h1>
-
-        <form onSubmit={this.handleSubmit}>
-          <input type="text" name="university" />
-          <button type="submit">
-            Add
-          </button>
-        </form>
+        <ListingForm afterSubmit={this.onAfterSubmit} />
       </div>
     );
   }
 }
 
-Page.propTypes = {
-  registryAddress: PropTypes.string,
-  participants: PropTypes.arrayOf(PropTypes.object),
-  submitApplication: PropTypes.func,
-};
-
-Page.defaultProps = {
-  registryAddress: '',
-  participants: [],
-  submitApplication: () => {},
-};
-
-const mapStateToProps = (state) => ({
-  registryAddress: getRegistryAddress(state),
-  participants: getParticipants(state),
-});
-
-const mapDispatchToProps = {
-  submitApplication: apply,
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(Page);
+export default IndexPage;
